Replaced unsafe `proc_address` function with `load_function`.
The overloads of `proc_address` that took a library name and function name were unsafe because they released their reference to the library before returning the function pointer.  If other code still had a handle to the library, the function would work because the library wouldn't be unloaded.  However, if the library was not previously loaded, the library would be unloaded as the reference is released, and calling the function would cause a crash.

This commit fixes that by replacing those overloads of `proc_address` with a function called `load_library` that returns a `boost::function` that holds a smart reference to the loaded library instead of returning a plain function pointer.  This way, the library will be held loaded at least until the function is out of scope and so can no longer be called.