Compile `task_dialog_builder` show method in test, even if not calling it.
Although we mostly don't want to call the show method, because we need the tests to remain automated, we should still be compiling it to make the test as useful as possible.  This change replaces comment out calls to `show()` with a conditional that will not, ordinarily, be true.  The condition also makes it easy to temporarily change the tests to show the dialogs.