Use threads safely in task dialog async update classes.
It is not safe to just detach the threads in our async update classes as they reference a function that is likely to use external state when executing.  This change using an RAII wrapper to join the threads before the last copy of the async update class goes out of scope.  Assuming the state used by the function outlives the async updater, this won't let the thread access invalid state.

http://www.open-std.org/Jtc1/sc22/wg21/docs/papers/2008/n2802.html
http://akrzemi1.wordpress.com/2012/11/14/not-using-stdthread/

We came accross this bug because the test suite main() was returning while the thread was still running.  This aborted the thread and leaked memory.