Dynamically load TaskDialogIndirect using correct calling convention.
When changing from `proc_address` to `load_function` (019e77d7), we accidentally removed the `WINAPI` (aka `__stdcall`) calling convention specifier.

This commit replaces that and changes `load_function` so that it strips the calling convention from the signature of the returned `boost::function` object.  This means the object can be called without the code needing knowledge of the underlying calling convention, which mirrors how `boost::function` works when passed a function pointer directly.