Renamed library from winapi to Washer.
The previous name, 'winapi' was accurate but useless for search engines because it is the common acronym for the Windows API (the thing we are wrapping).  We chose Washer as the new name because it was not already taken, and it links to the the word Windows (washing windows), and a washer is a layer you place between to objects when joining them together.